Move unlock balance modal to src
