+++ src/modules/dpos/validator/components/unlockBalanceForm/unlockBalance.css
@@ -1,4 +1,4 @@
-@import '../../../../../../../setup/react/app/mixins.css';
+@import '../../../../../../setup/react/app/mixins.css';
 
 .wrapper {
   & > header {
