+++ src/modules/dpos/validator/components/unlockBalanceSummary/unlockBalanceSummary.css
@@ -1,4 +1,4 @@
-@import '../../../../../../../setup/react/app/mixins.css';
+@import '../../../../../../setup/react/app/mixins.css';
 
 .summaryContainer {
   & > header {
