+++ src/modules/dpos/validator/components/unlockBalanceSummary/unlockBalanceSummary.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import wallets from '@tests/constants/wallets';
-import Summary from './LockedBalanceSummary';
+import Summary from './unlockBalanceSummary';
 
 describe('Locked balance Summary', () => {
   const props = {
