+++ src/modules/dpos/validator/components/unlockBalanceView/index.js
@@ -2,9 +2,9 @@ import React from 'react';
 import Dialog from 'src/theme/dialog/dialog';
 import MultiStep from 'src/modules/common/components/OldMultiStep';
 import TransactionSignature from '@transaction/components/TransactionSignature';
-import Form from './LockedBalanceForm';
-import Status from './LockedBalanceStatus';
-import Summary from './LockedBalanceSummary';
+import Form from '../unlockBalanceForm';
+import Status from '../unlockBalanceStatus';
+import Summary from '../unlockBalanceSummary';
 
 const Modal = () => (
   <Dialog hasClose>
